requirements.txt
ankh
torch
transformers
fair-esm